refactor(trusty-memory): extract chat cluster from web.rs to chat.rs
